+++ python_src/leb/freeze/simulation.py
@@ -7,18 +7,18 @@ from skimage.data import astronaut, camera
 from skimage.transform import resize
 
 from leb.freeze.calibration import Calibration, calibrate_rectangular_matrix
-from leb.freeze.datasets import PtychoDataset
-from leb.freeze.fp import Pupil, slice_fft
+from leb.freeze.datasets import FPDataset
+from leb.freeze.fp import Pupil, fp_recover, slice_fft
 
 
 def fp_simulation(
     gt_img_size: int = 256,
     scaling_factor: int = 4,
     px_size_um: float = 11,
-    wavelength_um: float = 0.488e-6,
+    wavelength_um: float = 0.488,
     mag: float = 10.0,
     na: float = 0.288,
-    num_leds: tuple[int, int] = (16, 16),
+    num_leds: tuple[int, int] = (5, 5),
     center_led: tuple[float, float] = (8, 8),
     led_pitch_mm: tuple[float, float] = (4, 4),
     axial_offset_mm: float = -50,
@@ -26,52 +26,77 @@ def fp_simulation(
     gt = ground_truth()
 
     # Compute the LED indexes
-    led_indexes = [(i, j) for i in range(num_leds[0]) for j in range(num_leds[1])]
+    led_indexes = generate_led_indexes(center_led, num_leds)
 
     # Get the wavevectors corresponding to the LEDs
-    led_pitch_um, axial_offset_um = led_pitch_mm * 1000, axial_offset_mm * 1000
-    wavevectors = calibrate_rectangular_matrix(
+    calibration = calibrate_rectangular_matrix(
         led_indexes,
         center_led,
-        pitch=led_pitch_um,
-        axial_offset=axial_offset_um,
-        wavelength=wavelength_um,
+        pitch_mm=led_pitch_mm,
+        axial_offset_mm=axial_offset_mm,
+        wavelength_um=wavelength_um,
+        sort=True,
     )
 
-    # Create the simulated pupil.
+    # Create the simulated pupil and images.
     # The images and the pupil will be scaling_factor times smaller than the ground truth in each
     # dimension.
     dataset_size = int(gt_img_size / scaling_factor)
     pupil = Pupil.from_system_params(
         num_px=dataset_size, px_size_um=px_size_um, wavelength_um=wavelength_um, mag=mag, na=na
     )
+    images = generate_simulated_images(gt, calibration, pupil)
+
+    # Create the dataset
+    dataset = FPDataset.from_calibration(
+        images=images,
+        calibration=calibration,
+    )
+
+    # Recover the ground truth
+    _ = fp_recover(dataset, pupil, upsampling_factor=scaling_factor)
+
+
+def generate_led_indexes(
+    center_led: tuple[int, int], num_leds: tuple[int, int]
+) -> list[tuple[int, int]]:
+    """Generates the LED indexes for a rectangular array of LEDs."""
+    low_x = center_led[0] - (num_leds[0]) // 2
+    high_x = center_led[0] + (num_leds[0]) // 2
+    low_y = center_led[1] - (num_leds[1]) // 2
+    high_y = center_led[1] + (num_leds[1]) // 2
+    led_indexes = [(i, j) for i in range(low_x, high_x) for j in range(low_y, high_y)]
+
+    return led_indexes
 
 
 def generate_simulated_images(
-    gt: NDArray[np.complex128], calibration: Calibration, pupil: Pupil, scaling_factor: int = 4
+    gt: NDArray[np.complex128], calibration: Calibration, pupil: Pupil
 ) -> NDArray[np.float64]:
     """Generates simulated images from a ground truth object."""
     gt_fft = fftshift(fft2(gt))
 
-    original_size_px = gt.shape[0]
-    rescaled_size_px = int(original_size_px / scaling_factor)
+    gt_size_px = gt.shape[0]
+    dataset_size_px = pupil.p.shape[0]
+    scaling_factor = gt_size_px // dataset_size_px
     num_images = len(calibration)
 
-    images = np.empty((num_images, rescaled_size_px, rescaled_size_px))
-    for image_num, led_indexes, wavevector in enumerate(calibration.items()):
-        # Obtain the rectangular slice from the FFT centered at kx, ky
+    images = np.empty((num_images, dataset_size_px, dataset_size_px))
+    for image_num, item in enumerate(calibration.items()):
+        _, wavevector = item
+        # Obtain the rectangular slice from the FFT centered at kx, ky.
+        # First convert (kx, ky) into pixels using a k-space spacing corresponding to the
+        # size of the groundtruth Fourier transform, i.e. pupil.dk / scaling_factor.
+        kx_ky_px = np.round(scaling_factor * np.array(wavevector[:2]) / pupil.dk).astype(int)
         current_slice_fft = slice_fft(
             gt_fft,
-            pupil,
-            np.array(wavevector),
-            scaling_factor,
-            original_size_px,
-            rescaled_size_px,
+            kx_ky_px,
+            dataset_size_px,
         )
 
         # Copy the slice, then low pass filter it with the pupil
         current_slice_fft = current_slice_fft.copy()
-        current_slice_fft *= pupil  # TODO Add abberrations here
+        current_slice_fft *= pupil.p  # TODO Add abberrations here
 
         # Inverse FFT to obtain the image
         current_image = pupil.dk**2 * np.abs(ifft2(ifftshift(current_slice_fft)))
@@ -114,3 +139,7 @@ def ground_truth(
     ground_truth = high_res_object * np.exp(1j * high_res_phase)
 
     return ground_truth
+
+
+if __name__ == "__main__":
+    fp_simulation()