+++ tests/leb-freeze/test_calibration.py
@@ -9,16 +9,16 @@ def test_calibrate_rectangular_matrix():
     """Test the rectangular matrix calibration for an on-axis LED."""
     indexes = [(0, 0)]
     center_led = (0, 0)
-    pitch = 4e3  # 4 mm
-    axial_offset = -50e3  # = 5 cm
-    wavelength = 0.488  # 488 nm
+    pitch_mm = 4
+    axial_offset_mm = -50
+    wavelength_um = 0.488
 
     ks = calibrate_rectangular_matrix(
-        indexes, center_led, pitch, axial_offset=axial_offset, wavelength=wavelength
+        indexes, center_led, pitch_mm, axial_offset_mm=axial_offset_mm, wavelength_um=wavelength_um
     )
 
     kx, ky, kz = ks[(0, 0)]
-    k = 2 * np.pi / wavelength
+    k = 2 * np.pi / wavelength_um
 
     assert_almost_equal(kx**2 + ky**2 + kz**2, k**2)
     assert_almost_equal(kx, 0)
@@ -30,18 +30,18 @@ def test_calibrate_rectangular_matrix_off_center():
     """Test the results of an off-axis LED."""
     indexes = [(0, 1)]
     center_led = (0, 0)
-    pitch = 4e3  # 4 mm
-    axial_offset = -50e3  # = 5 cm
-    wavelength = 0.488  # 488 nm
+    pitch_mm = 4
+    axial_offset_mm = -50
+    wavelength_um = 0.488
 
     ks = calibrate_rectangular_matrix(
-        indexes, center_led, pitch, axial_offset=axial_offset, wavelength=wavelength
+        indexes, center_led, pitch_mm, axial_offset_mm=axial_offset_mm, wavelength_um=wavelength_um
     )
     k_actual = np.array([*ks[(0, 1)]])
 
     # Expected wavevector
-    angle = np.arctan(pitch / axial_offset)
-    k = 2 * np.pi / wavelength
+    angle = np.arctan(pitch_mm / axial_offset_mm)
+    k = 2 * np.pi / wavelength_um
     k_expected = np.array([0, k * np.sin(angle), k * np.cos(angle)])
 
     assert_almost_equal(k_actual, k_expected)
@@ -51,24 +51,24 @@ def test_calibrate_rectangular_matrix_off_center_rotated():
     """Test the results of an off-axis LED with the matrix rotated 45 degrees from center."""
     indexes = [(0, 1)]
     center_led = (0, 0)
-    pitch = 4e3  # 4 mm
-    axial_offset = -50e3  # = 5 cm
+    pitch_mm = 4e3
+    axial_offset_mm = -50e3
     rot_deg = 45
-    wavelength = 0.488  # 488 nm
+    wavelength_um = 0.488
 
     ks = calibrate_rectangular_matrix(
         indexes,
         center_led,
-        pitch,
-        axial_offset=axial_offset,
+        pitch_mm,
+        axial_offset_mm=axial_offset_mm,
         rot_deg=rot_deg,
-        wavelength=wavelength,
+        wavelength_um=wavelength_um,
     )
     k_actual = np.array([*ks[(0, 1)]])
 
     # Expected wavevector
-    angle = np.arctan(pitch / axial_offset)
-    k = 2 * np.pi / wavelength
+    angle = np.arctan(pitch_mm / axial_offset_mm)
+    k = 2 * np.pi / wavelength_um
     k_expected = np.array(
         [np.sqrt(2) / 2 * k * np.sin(angle), np.sqrt(2) / 2 * k * np.sin(angle), k * np.cos(angle)]
     )
@@ -88,7 +88,12 @@ def test_calibrate_rectangular_matrix_sorted_results(sort, expected):
     wavelength = 0.488  # 488 nm
 
     ks = calibrate_rectangular_matrix(
-        indexes, center_led, pitch, axial_offset=axial_offset, wavelength=wavelength, sort=sort
+        indexes,
+        center_led,
+        pitch,
+        axial_offset_mm=axial_offset,
+        wavelength_um=wavelength,
+        sort=sort,
     )
 
     assert list(ks.keys()) == expected